feat: change port, fix dto

+++ src/CdrBilling.Application/DTOs/SubscriberSummaryDto.cs
@@ -3,6 +3,6 @@ namespace CdrBilling.Application.DTOs;
 public sealed record SubscriberSummaryDto(
     string PhoneNumber,
     string ClientName,
-    int CallCount,
-    int TotalBillableSec,
+    long CallCount,
+    long TotalBillableSec,
     decimal TotalCharge);
